[ListItems] Render tree should be all clean by the end of FrameView::layout(). https://bugs.webkit.org/show_bug.cgi?id=162832
Reviewed by Simon Fraser.

List item markers are inserted into the tree during layout, right before laying out
the list item (that's a big FIXME). At this point we already know what part of the tree
needs to be laid out. Inserting a list item marker does not expand this dirty area.
However whenever we insert a new renderer into the tree, we call setNeedsLayout on the contining block chain.
In certain cases (floating renderers), it could potentially trigger some unintentional markings
and we return from FrameView::layout() with a dirty subtree.

This patch preemptively marks the list item and its marker dirty so that
when the marker is getting inserted into the tree, we stop the marking at the parent.

Not testable.

* rendering/RenderListItem.cpp:
(WebCore::RenderListItem::insertOrMoveMarkerRendererIfNeeded):


Identifier: 180822@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@206765 268f45cc-cd09-0410-ab3c-d52691b4dbfc