REGRESSION (r207179): ASSERTION FAILED: node.cell != previousCell https://bugs.webkit.org/show_bug.cgi?id=163337
Reviewed by Mark Lam.

It turns out that HeapSnapshot was not down with revisiting. The concurrent GC is going to be
built around the idea that we can revisit objects many times. This means that any action that
should only take place once per object must check the object's state. This fixes the snapshot
code to do this.

While writing this code, I realized that we're actually doing this check incorrectly, so I
filed bug 163343. That bug requires a race, so we aren't going to see it yet.

* heap/HeapSnapshot.cpp:
(JSC::HeapSnapshot::finalize):
* heap/SlotVisitor.cpp:
(JSC::SlotVisitor::appendToMarkStack):
(JSC::SlotVisitor::visitChildren):



Identifier: 181178@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@207230 268f45cc-cd09-0410-ab3c-d52691b4dbfc