Invalid assertion in arguments elimination https://bugs.webkit.org/show_bug.cgi?id=163740 <rdar://problem/27911462>
Reviewed by Michael Saboff.

JSTests:

* stress/direct-arguments-osr-entry.js: Added.
(foo.):
(foo):

Source/JavaScriptCore:

The DFGFTL's arguments elimination phase incorrectly asserted that a GetFromArguments' first
child would always be a CreateDirectArguments.  While we only create the
op_get_from_arguments bytecode pointing to a create_direct_arguments, its possible for a
number of reasons that a DFG GetFromArguments may not point to a CreateDirectArguments. For
example, if we are OSR entering in some function with direct arguments the
CreateDirectArguments node might become ExtractOSREntryLocals.

* dfg/DFGArgumentsEliminationPhase.cpp:

Identifier: 181510@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@207623 268f45cc-cd09-0410-ab3c-d52691b4dbfc